+++ ApiaryDataCollector/Models/DTO/ApiaryDto.cs
@@ -5,7 +5,7 @@
         public int ApiaryNumber { get; set; }
         public DateTime ReportDate { get; set; }
         public int Humidity { get; set; }
-        public int Temperature { get; set; }
+        public double Temperature { get; set; }
         public int LightIntensity { get; set; }
         public List<HiveDto>? Hives { get; set; }
     }
